(refactor) DeferredCallback refactored by Aszepeshazi
